refactor reportstable
